# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API implementation.",
        "title": "Swagger System Payment",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/order/new": {
            "post": {
                "description": "save Order in database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Insert Order",
                "parameters": [
                    {
                        "description": "Order example",
                        "name": "example",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError500"
                        }
                    }
                }
            }
        },
        "/order/orders": {
            "get": {
                "description": "Select all Orders",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Select all Orders",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 0,
                        "description": "start example",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "description": "count example",
                        "name": "count",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.OrderResponse"
                            }
                        }
                    }
                }
            }
        },
        "/order/update": {
            "put": {
                "description": "Updates a Order in database (id req)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Updates Order",
                "parameters": [
                    {
                        "description": "Order example",
                        "name": "example",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError500"
                        }
                    }
                }
            }
        },
        "/order/{id}": {
            "get": {
                "description": "Get one Order from ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Select Order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "example: 1",
                        "name": "int",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError500"
                        }
                    }
                }
            }
        },
        "/payer/new": {
            "post": {
                "description": "save payer in database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payer"
                ],
                "summary": "Insert Payer",
                "parameters": [
                    {
                        "description": "Payer example",
                        "name": "example",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Payer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError500"
                        }
                    }
                }
            }
        },
        "/payer/payers": {
            "get": {
                "description": "Select all Payers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payer"
                ],
                "summary": "Select all Payers",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 0,
                        "description": "start example",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "description": "count example",
                        "name": "count",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.PayerResponse"
                            }
                        }
                    }
                }
            }
        },
        "/payer/update": {
            "put": {
                "description": "Updates a payer in database (id req)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payer"
                ],
                "summary": "Updates Payer",
                "parameters": [
                    {
                        "description": "Payer example",
                        "name": "example",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Payer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError500"
                        }
                    }
                }
            }
        },
        "/payer/{id}": {
            "get": {
                "description": "Get one Payer from ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payer"
                ],
                "summary": "Select Payer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "example: 1",
                        "name": "int",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httputil.HTTPError400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request body or query parameters"
                }
            }
        },
        "httputil.HTTPError404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "error": {
                    "type": "string",
                    "example": "Page not found"
                }
            }
        },
        "httputil.HTTPError500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "error": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
        "model.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "example": "Volta Redonda"
                },
                "number": {
                    "type": "string",
                    "example": "1106"
                },
                "state": {
                    "type": "string",
                    "example": "Rio de Janeiro"
                },
                "street": {
                    "type": "string",
                    "example": "Servidão B-1"
                },
                "zipCode": {
                    "type": "string",
                    "example": "27275-595"
                }
            }
        },
        "model.AddressResponse": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "example": "Volta Redonda"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "number": {
                    "type": "string",
                    "example": "1106"
                },
                "state": {
                    "type": "string",
                    "example": "Rio de Janeiro"
                },
                "street": {
                    "type": "string",
                    "example": "Servidão B-1"
                },
                "zip_code": {
                    "type": "string",
                    "example": "27275-595"
                }
            }
        },
        "model.Card": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string",
                    "example": "Visa"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "last4": {
                    "type": "string",
                    "example": "1234"
                },
                "payerID": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "model.Order": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "payerID": {
                    "type": "integer",
                    "example": 1
                },
                "productID": {
                    "type": "integer",
                    "example": 1
                },
                "totalFees": {
                    "type": "integer",
                    "maximum": 24,
                    "minimum": 1,
                    "example": 3
                }
            }
        },
        "model.OrderResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "finished": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "payer": {
                    "$ref": "#/definitions/model.PayerResponse"
                },
                "payments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PaymentResponse"
                    }
                },
                "product": {
                    "$ref": "#/definitions/model.Product"
                },
                "total_fees": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Payer": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/model.Address"
                },
                "birthDate": {
                    "type": "string",
                    "example": "24/07/1992"
                },
                "document": {
                    "type": "string",
                    "example": "23415162"
                },
                "email": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 6,
                    "example": "jhondoe@mail.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 6,
                    "example": "Jhon Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "+123456789"
                },
                "userReference": {
                    "type": "string",
                    "example": "12345"
                }
            }
        },
        "model.PayerResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/model.AddressResponse"
                },
                "birth_date": {
                    "type": "string",
                    "example": "24/07/1992"
                },
                "created_at": {
                    "type": "string"
                },
                "document": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "example": "jhondoe@mail.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Jhon Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "+123456789"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_reference": {
                    "type": "string",
                    "example": "12345"
                }
            }
        },
        "model.PaymentResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "card": {
                    "$ref": "#/definitions/model.Card"
                },
                "country": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "description": "Dlocal payment.id",
                    "type": "string"
                },
                "orderNumber": {
                    "description": "Dlocal order_id",
                    "type": "string"
                },
                "paymentMethodFlow": {
                    "type": "string"
                },
                "paymentMethodID": {
                    "type": "string"
                }
            }
        },
        "model.Product": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 5000
                },
                "currency": {
                    "type": "string",
                    "maxLength": 3,
                    "minLength": 3,
                    "example": "USD"
                },
                "description": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 6,
                    "example": "Curso de Programacion"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 6,
                    "example": "programacion en C"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": "\t\t\t\t\t\t\tGrants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": "\t\t\t\t\t\t\tGrants read and write access to administrative information",
                "write": "\t\t\t\t\t\t\tGrants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": "\t\t\t\t\t\t\tGrants read and write access to administrative information",
                "write": "\t\t\t\t\t\t\tGrants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
...
}